billion scale

+++ billion_scale_search/buildkNN.py
@@ -14,7 +14,7 @@ proportion = 0.1
 
 
 for i in range(int(1/proportion)):
-    processing_subset = np.ascontiguousarray(x[dataset_size*proportion*i : dataset_size*proportion*(i+1), :])
+    processing_subset = np.ascontiguousarray(x[int(dataset_size*proportion*i) : int(dataset_size*proportion*(i+1)), :])
     kmeans = faiss.Kmeans(d, k, niter = niter, verbose = True)
     kmeans.train(processing_subset)
     print(kmeans.obj)
